requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2
asyncpg
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
python-multipart